Blog List Expantion 4.18-4.23

+++ controllers/blogs.js
@@ -1,6 +1,7 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
@@ -8,17 +9,18 @@ blogsRouter.get('/', async (request, response) => {
     response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-    const body = request.body
-    const getUsers = await User.find({})
-    const user = getUsers[0]
+blogsRouter.post('/', userExtractor, async (request, response) => {
+    jwt.verify(request.token, process.env.SECRET)
+
+    const body = request.body    
+    const user = request.user
 
     const blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
         likes: body.likes,
-        user: user
+        user: user.id
     })
 
     const savedBlog = await blog.save()
@@ -29,12 +31,31 @@ blogsRouter.post('/', async (request, response) => {
     response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+    jwt.verify(request.token, process.env.SECRET)
+
+    const userId = request.user.id
+    const blogToDelete = await Blog.findById(request.params.id)
+
+    if (userId.toString() !== blogToDelete.user.toString()) {
+        return response.status(401).json({ error: 'invalid token' })
+    }
+
     await Blog.findByIdAndDelete(request.params.id)
+
     response.status(204).end()
 })
 
-blogsRouter.put('/:id', async (request, response) => {
+blogsRouter.put('/:id', userExtractor, async (request, response) => {
+    jwt.verify(request.token, process.env.SECRET)
+
+    const userId = request.user.id
+    const blogToUpdate = await Blog.findById(request.params.id)
+
+    if (userId.toString() !== blogToUpdate.user.toString()) {
+        return response.status(401).json({ error: 'invalid token' })
+    }
+
     const { title, author, url, likes } = request.body
 
     const updatedBlog = await Blog.findByIdAndUpdate(
